bytes refactored
