+++ CMakeLists.txt
@@ -123,7 +123,7 @@ set(PACKAGES
 	konoha.while
 	sugar
 	llvm
-	konoha.pcre
+##	konoha.pcre
 ##	konoha.io
 	posix.process
 	posix.file
@@ -131,7 +131,7 @@ set(PACKAGES
 ##	konoha.socket
 ##	konoha.rope
 ##	konoha.subproc
-	konoha.mpi
+##	konoha.mpi
 ##	curl
 ##	ijit
 ##	jansson
@@ -139,7 +139,7 @@ set(PACKAGES
 	python
 	apache
 #	openssl
-	logpool
+##	logpool
 )
 
 find_library(HAVE_LIBDL NAMES dl)
