+++ package/konoha.bytes/test/BytesDecodeFromEUCJP.k
@@ -2,9 +2,8 @@
 
 K.import("konoha.bytes");
 
-
 void test () {
-	 //p("'" + "ほげほげ".asBytes().encodeTo("EUC-JP").decodeFrom("EUC-JP") + "'");
+	 assert("hoge".asBytes().encodeTo("EUC-JP").decodeFrom("EUC-JP") == "hoge");
 	 assert("ほげほげ".asBytes().encodeTo("EUC-JP").decodeFrom("EUC-JP") == "ほげほげ");
 }
