+++ package/konoha.bytes/test/BytesEncodeToEUCJP.k.proof
@@ -1,0 +1,2 @@
+ - (BytesEncodeToEUCJP.k:7) ほげほげ
+Q.E.D.
