+++ package/konoha.bytes/test/BytesEncodeToShiftJIS.k.proof
@@ -1,0 +1,2 @@
+ - (BytesEncodeToShiftJIS.k:7) ‚Ù‚°‚Ù‚°
+Q.E.D.
